Port AWSClusterReconciler to kubebuilder v2
