--- pkg/controller/awscluster/awscluster_controller.go
@@ -1,234 +1,0 @@
-/*
-Copyright 2019 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package awscluster
-
-import (
-	"context"
-	"fmt"
-	"time"
-
-	"sigs.k8s.io/cluster-api-provider-aws/pkg/cloud/services/ec2"
-	"sigs.k8s.io/cluster-api-provider-aws/pkg/cloud/services/elb"
-
-	"sigs.k8s.io/cluster-api/pkg/util"
-
-	"github.com/pkg/errors"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/client-go/tools/record"
-	infrav1 "sigs.k8s.io/cluster-api-provider-aws/api/v1alpha2"
-	"sigs.k8s.io/cluster-api-provider-aws/pkg/cloud/scope"
-	"sigs.k8s.io/cluster-api/api/v1alpha2"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/controller"
-	"sigs.k8s.io/controller-runtime/pkg/handler"
-	"sigs.k8s.io/controller-runtime/pkg/log"
-	"sigs.k8s.io/controller-runtime/pkg/manager"
-	"sigs.k8s.io/controller-runtime/pkg/reconcile"
-	"sigs.k8s.io/controller-runtime/pkg/source"
-)
-
-const (
-	controllerName  = "awscluster-controller"
-	apiEndpointPort = 6443
-)
-
-// Add creates a new AWSCluster Controller and adds it to the Manager with default RBAC. The Manager will set fields on the Controller
-// and Start it when the Manager is Started.
-func Add(mgr manager.Manager) error {
-	return add(mgr, newReconciler(mgr))
-}
-
-// newReconciler returns a new reconcile.Reconciler
-func newReconciler(mgr manager.Manager) reconcile.Reconciler {
-	return &ReconcileAWSCluster{
-		Client:   mgr.GetClient(),
-		scheme:   mgr.GetScheme(),
-		recorder: mgr.GetEventRecorderFor(controllerName),
-	}
-}
-
-// add adds a new Controller to mgr with r as the reconcile.Reconciler
-func add(mgr manager.Manager, r reconcile.Reconciler) error {
-	// Create a new controller
-	c, err := controller.New(controllerName, mgr, controller.Options{Reconciler: r})
-	if err != nil {
-		return err
-	}
-
-	// Watch for changes to AWSCluster
-	err = c.Watch(
-		&source.Kind{Type: &infrav1.AWSCluster{}},
-		&handler.EnqueueRequestForObject{},
-	)
-	if err != nil {
-		return err
-	}
-
-	// Watch for changes to Cluster and enqueue the associated AWSCluster
-	return c.Watch(
-		&source.Kind{Type: &v1alpha2.Cluster{}},
-		&handler.EnqueueRequestsFromMapFunc{
-			ToRequests: util.ClusterToInfrastructureMapFunc(schema.GroupVersionKind{
-				Group:   infrav1.SchemeBuilder.GroupVersion.Group,
-				Version: infrav1.SchemeBuilder.GroupVersion.Version,
-				Kind:    "AWSCluster",
-			}),
-		},
-	)
-}
-
-var _ reconcile.Reconciler = &ReconcileAWSCluster{}
-
-// ReconcileAWSCluster reconciles a AWSCluster object
-type ReconcileAWSCluster struct {
-	client.Client
-	scheme   *runtime.Scheme
-	recorder record.EventRecorder
-}
-
-// Reconcile reads that state of the cluster for a AWSCluster object and makes changes based on the state read
-// and what is in the AWSCluster.Spec
-func (r *ReconcileAWSCluster) Reconcile(request reconcile.Request) (_ reconcile.Result, reterr error) {
-	ctx := context.TODO()
-	logger := log.Log.WithName(controllerName).
-		WithName(fmt.Sprintf("namespace=%s", request.Namespace)).
-		WithName(fmt.Sprintf("awsCluster=%s", request.Name))
-
-	// Fetch the AWSCluster instance
-	awsCluster := &infrav1.AWSCluster{}
-	err := r.Get(ctx, request.NamespacedName, awsCluster)
-	if err != nil {
-		if apierrors.IsNotFound(err) {
-			return reconcile.Result{}, nil
-		}
-		return reconcile.Result{}, err
-	}
-
-	logger = logger.WithName(awsCluster.APIVersion)
-
-	// Fetch the Cluster.
-	cluster, err := util.GetOwnerCluster(ctx, r.Client, awsCluster.ObjectMeta)
-	if err != nil {
-		return reconcile.Result{}, err
-	}
-	if cluster == nil {
-		logger.Info("Waiting for Cluster Controller to set OwnerRef on AWSCluster")
-		return reconcile.Result{RequeueAfter: 10 * time.Second}, nil
-	}
-
-	logger = logger.WithName(fmt.Sprintf("cluster=%s", cluster.Name))
-
-	// Create the scope.
-	clusterScope, err := scope.NewClusterScope(scope.ClusterScopeParams{
-		Client:     r.Client,
-		Logger:     logger,
-		Cluster:    cluster,
-		AWSCluster: awsCluster,
-	})
-	if err != nil {
-		return reconcile.Result{}, errors.Errorf("failed to create scope: %+v", err)
-	}
-
-	// Always close the scope when exiting this function so we can persist any AWSMachine changes.
-	defer func() {
-		if err := clusterScope.Close(); err != nil && reterr == nil {
-			reterr = err
-		}
-	}()
-
-	// Handle deleted clusters
-	if !awsCluster.DeletionTimestamp.IsZero() {
-		return reconcileDelete(clusterScope)
-	}
-
-	// Handle non-deleted clusters
-	return reconcileNormal(clusterScope)
-}
-
-// TODO(ncdc): should this be a function on ClusterScope?
-func reconcileDelete(clusterScope *scope.ClusterScope) (reconcile.Result, error) {
-	clusterScope.Info("Reconciling AWSCluster delete")
-
-	ec2svc := ec2.NewService(clusterScope)
-	elbsvc := elb.NewService(clusterScope)
-	awsCluster := clusterScope.AWSCluster
-
-	if err := elbsvc.DeleteLoadbalancers(); err != nil {
-		return reconcile.Result{}, errors.Wrapf(err, "error deleting load balancer for AWSCluster %s/%s", awsCluster.Namespace, awsCluster.Name)
-	}
-
-	if err := ec2svc.DeleteBastion(); err != nil {
-		return reconcile.Result{}, errors.Wrapf(err, "error deleting bastion for AWSCluster %s/%s", awsCluster.Namespace, awsCluster.Name)
-	}
-
-	if err := ec2svc.DeleteNetwork(); err != nil {
-		return reconcile.Result{}, errors.Wrapf(err, "error deleting network for AWSCluster %s/%s", awsCluster.Namespace, awsCluster.Name)
-	}
-
-	// Cluster is deleted so remove the finalizer.
-	clusterScope.AWSCluster.Finalizers = util.Filter(clusterScope.AWSCluster.Finalizers, infrav1.ClusterFinalizer)
-
-	return reconcile.Result{}, nil
-}
-
-// TODO(ncdc): should this be a function on ClusterScope?
-func reconcileNormal(clusterScope *scope.ClusterScope) (reconcile.Result, error) {
-	clusterScope.Info("Reconciling AWSCluster")
-
-	awsCluster := clusterScope.AWSCluster
-
-	// If the AWSCluster doesn't have our finalizer, add it.
-	if !util.Contains(awsCluster.Finalizers, infrav1.ClusterFinalizer) {
-		awsCluster.Finalizers = append(awsCluster.Finalizers, infrav1.ClusterFinalizer)
-	}
-
-	ec2Service := ec2.NewService(clusterScope)
-	elbService := elb.NewService(clusterScope)
-
-	if err := ec2Service.ReconcileNetwork(); err != nil {
-		return reconcile.Result{}, errors.Wrapf(err, "failed to reconcile network for AWSCluster %s/%s", awsCluster.Namespace, awsCluster.Name)
-	}
-
-	if err := ec2Service.ReconcileBastion(); err != nil {
-		return reconcile.Result{}, errors.Wrapf(err, "failed to reconcile bastion host for AWSCluster %s/%s", awsCluster.Namespace, awsCluster.Name)
-	}
-
-	if err := elbService.ReconcileLoadbalancers(); err != nil {
-		return reconcile.Result{}, errors.Wrapf(err, "failed to reconcile load balancers for AWSCluster %s/%s", awsCluster.Namespace, awsCluster.Name)
-	}
-
-	if awsCluster.Status.Network.APIServerELB.DNSName == "" {
-		clusterScope.Info("Waiting on API server ELB DNS name")
-		return reconcile.Result{RequeueAfter: 15 * time.Second}, nil
-	}
-
-	// Set APIEndpoints so the Cluster API Cluster Controller can pull them
-	awsCluster.Status.APIEndpoints = []infrav1.APIEndpoint{
-		{
-			Host: awsCluster.Status.Network.APIServerELB.DNSName,
-			// TODO(ncdc): should this come from awsCluster.Status.Network.APIServerELB.Listeners[0].Port?
-			Port: apiEndpointPort,
-		},
-	}
-
-	// No errors, so mark us ready so the Cluster API Cluster Controller can pull it
-	awsCluster.Status.Ready = true
-
-	return reconcile.Result{}, nil
-}
